requirements.txt
Flask==3.1.0
waitress==3.0.2
SQLAlchemy==2.0.24
python-dotenv==1.0.0